requirements.txt
customtkinter>=5.2.0
pyyaml>=6.0.1